+++ __snapshots__/packages/java-edition/test-out/mcfunction/parser/argument/minecraftComponent.spec.js
@@ -263,7 +263,11 @@ exports['mcfunction argument minecraft:component Parse "{"text":"hello world"}" 
               "quote": "\""
             }
           }
-        ]
+        ],
+        "innerRange": {
+          "start": 1,
+          "end": 21
+        }
       }
     ],
     "targetType": {