+++ packages/mcdoc/src/binder/index.ts
@@ -54,7 +54,6 @@ import {
 	ListTypeNode,
 	LiteralNode,
 	LiteralTypeNode,
-	LongRangeNode,
 	NumericTypeNode,
 	PathNode,
 	PrimitiveArrayTypeNode,
@@ -885,7 +884,7 @@ function convertEnumField(node: EnumFieldNode, ctx: McdocBinderContext): EnumTyp
 	}
 }
 
-function convertEnumValue(node: EnumValueNode, ctx: McdocBinderContext): string | number | bigint {
+function convertEnumValue(node: EnumValueNode, ctx: McdocBinderContext): string | number {
 	if (TypedNumberNode.is(node)) {
 		const { value } = TypedNumberNode.destruct(node)
 		return value.value
@@ -987,30 +986,23 @@ function convertList(node: ListTypeNode, ctx: McdocBinderContext): McdocType {
 	return wrapType(node, {
 		kind: 'list',
 		item: convertType(item, ctx),
-		lengthRange: convertRange(lengthRange),
+		lengthRange: convertRange(lengthRange, ctx),
 	}, ctx)
 }
 
-function convertRange(node: FloatRangeNode | IntRangeNode): NumericRange<number>
-function convertRange(node: LongRangeNode): NumericRange<bigint>
+function convertRange(node: FloatRangeNode | IntRangeNode, ctx: McdocBinderContext): NumericRange
 function convertRange(
 	node: FloatRangeNode | IntRangeNode | undefined,
-): NumericRange<number> | undefined
-function convertRange(node: LongRangeNode | undefined): NumericRange<bigint> | undefined
+	ctx: McdocBinderContext,
+): NumericRange | undefined
 function convertRange(
-	node: FloatRangeNode | IntRangeNode | LongRangeNode | undefined,
-): NumericRange<number> | NumericRange<bigint> | undefined
-function convertRange(
-	node: FloatRangeNode | IntRangeNode | LongRangeNode | undefined,
-): NumericRange | NumericRange<bigint> | undefined {
+	node: FloatRangeNode | IntRangeNode | undefined,
+	ctx: McdocBinderContext,
+): NumericRange | undefined {
 	if (!node) {
 		return undefined
 	}
 
-	if (LongRangeNode.is(node)) {
-		const { kind, min, max } = LongRangeNode.destruct(node)
-		return { kind, min: min?.value, max: max?.value }
-	}
 	const { kind, min, max } = FloatRangeNode.is(node)
 		? FloatRangeNode.destruct(node)
 		: IntRangeNode.destruct(node)
@@ -1031,7 +1023,7 @@ function convertLiteralValue(node: LiteralTypeValueNode, ctx: McdocBinderContext
 			kind: convertLiteralNumberSuffix(suffix, ctx)
 				?? (value.type === 'integer' ? 'int' : 'double'),
 			value: value.value,
-		} as LiteralValue
+		}
 	} else {
 		return { kind: 'string', value: node.value }
 	}
@@ -1062,22 +1054,22 @@ function convertNumericType(node: NumericTypeNode, ctx: McdocBinderContext): Mcd
 	const { numericKind, valueRange } = NumericTypeNode.destruct(node)
 	return wrapType(node, {
 		kind: numericKind.value as NumericTypeKind,
-		valueRange: convertRange(valueRange),
-	} as McdocType, ctx)
+		valueRange: convertRange(valueRange, ctx),
+	}, ctx)
 }
 
 function convertPrimitiveArray(node: PrimitiveArrayTypeNode, ctx: McdocBinderContext): McdocType {
 	const { arrayKind, lengthRange, valueRange } = PrimitiveArrayTypeNode.destruct(node)
 	return wrapType(node, {
 		kind: `${arrayKind.value as PrimitiveArrayValueKind}_array`,
-		lengthRange: convertRange(lengthRange),
-		valueRange: convertRange(valueRange),
-	} as McdocType, ctx)
+		lengthRange: convertRange(lengthRange, ctx),
+		valueRange: convertRange(valueRange, ctx),
+	}, ctx)
 }
 
 function convertString(node: StringTypeNode, ctx: McdocBinderContext): McdocType {
 	const { lengthRange } = StringTypeNode.destruct(node)
-	return wrapType(node, { kind: 'string', lengthRange: convertRange(lengthRange) }, ctx)
+	return wrapType(node, { kind: 'string', lengthRange: convertRange(lengthRange, ctx) }, ctx)
 }
 
 function convertReference(node: ReferenceTypeNode, ctx: McdocBinderContext): McdocType {