+++ packages/nbt/src/checker/index.ts
@@ -13,7 +13,7 @@ import {
 } from '../node/index.js'
 import { getBlocksFromItem, getEntityFromItem } from './mcdocUtil.js'
 
-const typed: core.Checker<TypedNbtNode> = (node, ctx) => {
+export const typed: core.SyncChecker<TypedNbtNode> = (node, ctx) => {
 	typeDefinition(node.targetType)(node.children[0], ctx)
 }
 
@@ -195,7 +195,11 @@ function inferType(node: NbtNode): Exclude<mcdoc.McdocType, mcdoc.UnionType> {
 		case 'nbt:float':
 			return { kind: 'literal', value: { kind: 'float', value: node.value } }
 		case 'nbt:long':
-			return { kind: 'literal', value: { kind: 'long', value: node.value } }
+			return {
+				kind: 'literal',
+				// TODO: this should NOT change type from `bigint` to `number`
+				value: { kind: 'long', value: Number(node.value) },
+			}
 		case 'nbt:int':
 			return { kind: 'literal', value: { kind: 'int', value: node.value } }
 		case 'nbt:short':